requirements.txt
django~=4.2.22
djangorestframework~=3.16.0
djangorestframework-simplejwt~=5.5.0